put the motion.py and path_planing.py together cause they both control the human in this program
